Create Dog3.java
When a method has a property, that's to say, we set up the classification for all the samples and later on pick one specific sample from a whole class, the method itself is no longer "static". So we cannot write "static" in this method. However, in the main function, we keep the "static" because that applies to only one dog with a specific weightinpounds.